Attempt to rescue some old Korean comments.
It appears like some old comments in Korean were lost in an
old commit, and in particular OOo-SVN253705. An attempt
to rescue the comments from the old SUN code was made
along with a computer aided translation which is much
more useful and less sensitive to break again.

No functional change is expected. 


git-svn-id: https://svn.apache.org/repos/asf/incubator/ooo/trunk@1333151 13f79535-47bb-0310-9956-ffa450edef68