Check svn info output for Linux Subversion
In current Linux (and cygwin) Subversion, the error message when there
is no repository on the current directory is "svn: '.' is not a working
copy".

NOTE: this message can be localized, so this only a temporary solution.
The code should really check the `svn info` return code, not the output.

git-svn-id: https://svn.apache.org/repos/asf/incubator/ooo/trunk@1341708 13f79535-47bb-0310-9956-ffa450edef68