i119384 - Python: replace deprecated compiler module with ast
The "compiler" module is deprecated on Python 2.6 and was
removed on 3.x so it's a good time to replace it. The
ast module was introduced in Python 2.5 so this will cause
trouble if using very old versions of system python.

While here run the python reindent tool on the adjacent
files.

Author:	hanya 


git-svn-id: https://svn.apache.org/repos/asf/incubator/ooo/trunk@1366506 13f79535-47bb-0310-9956-ffa450edef68