allow autotests to run on a DISPLAY other than :0.0
Running the automated tests on another display is a good idea, e.g. because our
autotests currently depend on the X11 keyboard being set to an en_US layout
(as defined in test/testcommon/source/org/openoffice/test/vcl/Tester.java),
to allow multiple test sessions and to allow running the tests on a developer
machine without disrupting the normal work too much. Being able to set the
DISPLAY variable to be different from :0.0 is very useful for all this.

FWIW: an X11 server that works well with our automated tests is
    Xephyr :99 -keybd ephyr,,,xkbmodel=evdev,xkblayout=en_US -screen 1440x900
and the DISPLAY variable has to be set to :99 for running with that.


git-svn-id: https://svn.apache.org/repos/asf/incubator/ooo/trunk@1396001 13f79535-47bb-0310-9956-ffa450edef68