Clean up the python detection for Python 3.
While here request Python be at least version 2.3. While this
version is likely to cause trouble, there is no other option
for older Mac PPC.

The ongoing updates in the python support will require bumping
the minimum to at least Python 2.6 but the recommendation is to
use the internal Python 2.7.x.


git-svn-id: https://svn.apache.org/repos/asf/incubator/ooo/trunk@1399479 13f79535-47bb-0310-9956-ffa450edef68