allow building an unchanged binfilter after changes to common header files
Cleanups in header files can result in finding or avoiding errors,
but the binfilter module as a legacy module should not be changed to
adjust for such cleanups.

In order to allow that the binfilter module can stay unaffected
by such changes the compiler macro BINFILTER_COMPAT has been added
to allow cleanups without disturbing binfilter.

When the flag is active header files are compatible to binfilter assumptions.
The flag is for now only active when compiling the binfilter module.

Having a grepable define allows further cleanups when the binfilter
module has been removed for good in a future release.


git-svn-id: https://svn.apache.org/repos/asf/incubator/ooo/trunk@1400130 13f79535-47bb-0310-9956-ffa450edef68