Bump SV_COUNTRY_LANGUAGE_OFFSET - Max count of formats per country/language
The previous value was 5000, we are bumping it to 8000. This is rather 
arbitrary number based on a similar change done in LibreOffice by Eike 
Rathke.

According to the Eike's commit:

Due to an insane amount of ridiculous "user-defined" number formats the
internal table may overflow during import, stave off and increase the
number of available slots to another arbitrary limit.


git-svn-id: https://svn.apache.org/repos/asf/incubator/ooo/trunk/main@1413213 13f79535-47bb-0310-9956-ffa450edef68