i121421 - Calc's RAND() behaves poorly on most platforms.
The random number generation function from libc was useful for very basic 
purposes but hasn't kept up with the times.

There are many options to provide a Portable Random Number Generator: one 
of the most popular and easiest to implement was designed by B.A Wichmann
and I.D Hill in 1982.

We chose to implement the newer 2006[1] algorithm which is better suited
modern platforms and passes DIEHARD and TestU01 - Big Crush tests.

Use of libc's rand() has been completely eliminated by using preexisting
rtl/random support to generate the initial seeds.

Special thanks to orcmid@ for discussion and testing.

Reference:

[1] B. A. Wichmann and Hill, Generating good pseudorandom numbers,
    Computational Statistics & Data Analysis, Volume 51 Issue 3,
    December, 2006, Pages 1614-1622.


git-svn-id: https://svn.apache.org/repos/asf/openoffice/trunk@1416271 13f79535-47bb-0310-9956-ffa450edef68