#i121442# - Allow redrawing a single status bar item
A StatusbarController that draws its status bar item may need to trigger
a repainting of its item when the feature state changes. Currently, the
only way to trigger this redraw is by changing the item text or the item
data, so that the vcl implementation calls ImplDrawItem; this is rather
implementation-specific.

git-svn-id: https://svn.apache.org/repos/asf/openoffice/trunk@1417846 13f79535-47bb-0310-9956-ffa450edef68