#i121442# - API wrapping StatusBar functionality
In the current implementation and its API design, it is impossible to
implement a StatusbarController in UNO, because there is no way in UNO
for the controller to manipulate its status bar item; the current
implementation of UNO based StatusbarController have to access directly
the VCL StatusBar through the AWT XWindow in order to manipulate its
item.

Designing an API that wraps the VCL StatusBar does not make sense: a
StatusbarController should only have access to the status bar item it
controls, not the whole status bar. Therefore this new interface
represents only the functionality of an item in a status bar, to be
controlled by a StatusbarController.

git-svn-id: https://svn.apache.org/repos/asf/openoffice/trunk@1417849 13f79535-47bb-0310-9956-ffa450edef68