i114430 - AOO Calc: 0^0=1 and it should be indeterminate.
Many spreadsheets and math packages set the value of POWER( 0, 0)
to 1 as that sometimes is required for some calculations. This
is also done by the standard C99 libc pow() function.

In mathematical terms the value is an indeterminate form and
in strict mathematical sense such value should at least be
considered with caution.

OpenFormula permits implementation defined values of 0, 1
or error in this case. We shall now return NaN.

Clean some headers while here.


git-svn-id: https://svn.apache.org/repos/asf/openoffice/trunk@1444110 13f79535-47bb-0310-9956-ffa450edef68