A miscellaneous use of boost.
Sinc(x) is a simple function used in communications. boost
happens to use it for quaternion math. The implementation is
simple, however it looks fun to re-use the boost function
even if just to say we are using quaternion math somewhere ;).

The performance difference is not likely to be huge but JIC,
set up a specific Boost math policy to limit type promotion
with it's corresponding impact. 


git-svn-id: https://svn.apache.org/repos/asf/openoffice/trunk@1444115 13f79535-47bb-0310-9956-ffa450edef68