fix vigra patch for BasicImage::resizeCopy()
"data_" is an actual data pointer but "data" is a member function
(since vigra-1.4). The method resizeCopy() seems to confuse data and data_
which the clang compiler was happy to point out. The OOo patch to vigra-1.4
already solved that problem but it was lost in the upgrade to vigra-1.6


git-svn-id: https://svn.apache.org/repos/asf/openoffice/trunk/main@1444805 13f79535-47bb-0310-9956-ffa450edef68