get more detailed info about the build platform's release and version
On the mailing lists there are frequently questions about the binary compatibility
of provided builds, e.g. from the buildbots. Autoconf and configure are bit skimpy
on the details of the build platform so the questions cannot be easily answered.

This change tries to get at least the details provided by /etc/*release and uname.


git-svn-id: https://svn.apache.org/repos/asf/openoffice/trunk@1452693 13f79535-47bb-0310-9956-ffa450edef68