avoid problems with C++ initializer lists (N2672) in our Aqua code
when the provided type and the target type of an initializer list don't match
the new behaviour described in C++11's N2672 proposal can be triggered. Though
that is very useful in general it complicates things for our case so that we're
better of to use the little helpers provided by XCode's CF framework.


git-svn-id: https://svn.apache.org/repos/asf/openoffice/trunk@1468907 13f79535-47bb-0310-9956-ffa450edef68