#i121715# add a CxxAbi compatible UNO bridge for 64bit MacOSX
the calling conventions on 64bit Mac follow the sysv-abi for AMD64 like
the FreeBSD and the Linux ports, so basing this bridge on the one from
FreeBSD saved some time. Thanks to Pedro Giffuni for this suggestion!

Some modififications were needed regarding symbol names, the layout of
unwind and exception structures, the missing derivatives of type_info
and vtable-offsets. The eh_frame specific stuff is not converted yet.


git-svn-id: https://svn.apache.org/repos/asf/openoffice/trunk@1476927 13f79535-47bb-0310-9956-ffa450edef68