#i122378# avoid non-iterator bound std::transform() in namedvaluecollection.hxx
in some build environments the use of std::transform() with plain pointers as
input iterators results in many warnings about how unsafe such a construct is.
The warnings could be suppressed e.g. on MSVC with the SCL_SECURE_NO_WARNINGS
define. Open coding the construct makes it cleaner and more debugable though.


git-svn-id: https://svn.apache.org/repos/asf/openoffice/trunk/main@1486840 13f79535-47bb-0310-9956-ffa450edef68