#i122603# fix database wizard's use of the Numberformatter
Patch by: Tsutomu Uchino <hanya.runo@gmail.com>
Review by: Herbert Durr <hdu@apache.org>

When the setNumberFormat method is called from DBColumn::initializeNumberFormat
method the number formatter is the document's one and the passed arguments
are from the database column.

Before the format key is set to the table cell, the format key from the 
column is used to generate the new key on the number format supplier. 
Therefore setting the cell's FormatsSupplier property is not required 
as the document's one is used.

As the same setNumberFormat method is called from other wizards the line 
is not removed but only active if the property exists.


git-svn-id: https://svn.apache.org/repos/asf/openoffice/trunk@1501137 13f79535-47bb-0310-9956-ffa450edef68