#i123142# handle "file changed by others" msgbox in autotests
autotests are routinely deleting/rewriting/overwriting temporary files
and a warning msgbox pops up for that which needs to be handled, else
the autotest runs into a timeout and thus fails.

Unfortunately the "file changed by others" message box doesn't have
an e.g. helpid or other constant for finding it, so the msgbox's
message itself has to be checked.


git-svn-id: https://svn.apache.org/repos/asf/openoffice/trunk@1518666 13f79535-47bb-0310-9956-ffa450edef68