#i123160# fix race-condition in fvt.gui.sd.shape.ShapeTypes
The two fvt.gui.sd.shape.ShapeTypes tests (for callout and star shapes)
often fail because the tests expects three items in the navigator as a
precondition. These three items are only shown there if filter->shapes
and then all-shapes has been selected.

There is a race condition: The click on filter->shapes fails when the
window-manager hasn't focussed the Navigator window. Then the precondition
that all shapes are selected fails.


git-svn-id: https://svn.apache.org/repos/asf/openoffice/trunk@1519425 13f79535-47bb-0310-9956-ffa450edef68