#i122195# adjust bridge for jsc's cxx->s5abi rename for OSX 64bit
System V AMD64 is the standard ABI on this platform. Every C++ compiler targets
it, so the generic "cxx" name was used previously for that ABI target. After
jsc's renaming COMID from cxx to s5abi the corresponding bridge directory name
has to be adjusted to the new COMID name too for consistency reasons.


git-svn-id: https://svn.apache.org/repos/asf/openoffice/branches/rejuvenate01@1534098 13f79535-47bb-0310-9956-ffa450edef68