#i123623# RmMoz 8/9: remove the obsoleted seamonkey 1.1.14 module
That module was used for its security libs and its address book support.
With #i122365# the dependency on its security libs was replaced with plain NSS.
With #i123575# the already quite bit-rotten seamonkey address books were removed.
Now there is nothing left for it to do and so that huge and complex module can
finally be retired after all these years.


git-svn-id: https://svn.apache.org/repos/asf/openoffice/trunk/main@1539267 13f79535-47bb-0310-9956-ffa450edef68