#i122208# replace the binaryurp cache for improved C++ compatibility
The C++ standards allows that the instantiation of incomplete types fails. The
Map::iterator in binaryurp Cache's Entry members had this problem. This rewrite
makes the code work with all compliant C++ compilers/STLs such as clang/libc++.
A Cache variant using an unordered_map is also provided and may be faster.
An interesting alternative would be to use boost's multi_index_container.


git-svn-id: https://svn.apache.org/repos/asf/openoffice/trunk/main@1550375 13f79535-47bb-0310-9956-ffa450edef68