#i123947# fix stlport emulation wrappers when compiling in C++11 mode
When compiling in C++11 mode the system provided TR1 compliant headers are
already required to be available and to be in the std namespace. The wrappers
trying to map the TR1 containers into the std namespace too is problematic
them, so it must be avoided in C++11 mode. Having TR1 compliant headers
available directly also simplifies the stlport4 emulation considerably.


git-svn-id: https://svn.apache.org/repos/asf/openoffice/trunk@1554812 13f79535-47bb-0310-9956-ffa450edef68