#i124421# force the availability of type_info symbols thrown by the ucb module
When exceptions are passed around in uno::Any types to be thrown later it
was observed on OSX64 that the type_info symbols of these exceptions were
not available anywhere. On most platforms the UNO bridge would synthesize
them by emulating how the platform's toolchain would do it. Forcing the
emission of the type_infos by the real toolchain is much better than
second guessing and reverse engineering its operation.

An even better approach would be to have the ucb throw the exceptions itself
instead of wrapping, rewrapping, unwrapping, etc. them. Once this is done
the dummyThrower helper function is no longer needed.


git-svn-id: https://svn.apache.org/repos/asf/openoffice/trunk@1582365 13f79535-47bb-0310-9956-ffa450edef68