#i124394# fix inclusion of freetype's ftsynth.h header
the location of this header file changed with freetype 2.51 so we had to adapt.
On the other hand the header is not actually needed because we load the symbols
for synthetization of outlines dynamically to be compatible with freetype libs
that don't support them. So the ftsynth.h header doesn't need inclusion at all.


git-svn-id: https://svn.apache.org/repos/asf/openoffice/branches/AOO410@1585412 13f79535-47bb-0310-9956-ffa450edef68