#i124617# make CoreText tolerate PDF-export's unexpected font switches
For some documents it was observed that the PDF-export switches the font after
text layout and before text drawing/measuring. This use case is quite atrocious
and unexpected and crashed our CoreText glue code. The other platforms survive
it though; to minimize the risk to them only the CoreText glue gets adapted to
survive this scenario.

Merged from trunk revision 1586249.


git-svn-id: https://svn.apache.org/repos/asf/openoffice/branches/AOO410@1586250 13f79535-47bb-0310-9956-ffa450edef68