Many spelling fixes: directories a* - g*.
Attempt to clean up most, but certainly not all, the spelling
mistakes that found home in OpenOffice through decades. We
could probably blame the international nature of the code but
it is somewhat shameful that this wasn't done before.

This is somewhat superfluous but surely more important than
translating German comments.


git-svn-id: https://svn.apache.org/repos/asf/openoffice/trunk/main@1591058 13f79535-47bb-0310-9956-ffa450edef68