#i124875# use AVAsset's best guess duration if AVPlayerItem isn't ready yet
in AV Foundation the most accurate media asset duration is available when the
AVPlayerItem's status changes to AVPlayerItemStatusReadyToPlay. There already
is an observer for this, but the upper layers expect getDuration() to be
synchronous and to return within a reasonable time. Since returning an
approximation is sufficient using the AVAsset's duration as fallback is fine.


git-svn-id: https://svn.apache.org/repos/asf/openoffice/trunk@1594838 13f79535-47bb-0310-9956-ffa450edef68
(cherry picked from commit 020bbb1d88621b9ceb9fc79a93df7017050e38c6)