#i124896# remove obsoleted sal-module-internal custom STL allocator
The SAL module avoided the heavy dependency on the stlport4-libraries by using
a custom allocator for its internal STL containers. With stlport4 removed these
dependencies are gone and the SAL-internal custom allocator is obsoleted. Since
the custom allocator results in build problems with clang>=3.4 or xcode>=5.1 it
is time to remove it for good.


git-svn-id: https://svn.apache.org/repos/asf/openoffice/trunk/main@1594842 13f79535-47bb-0310-9956-ffa450edef68