#i125020# fix rounding error in AquaSalGraphics::RefreshRect()
Rounding left and width down can accumulate a rounding error of almost two in
the calculation of right. The existing code compensates for this by increasing
the width by two, but since left has been decremented by one to accommodate for
antialiasing artifacts this compensation is one to little. Y-pos and height have
the same problems and get the matching fix.


git-svn-id: https://svn.apache.org/repos/asf/openoffice/trunk@1599173 13f79535-47bb-0310-9956-ffa450edef68