Fix webdav build with external serf.
Building with external serf (which is recommended) has been
broken for a while. Restore the proper include paths for
the external serf to work.

While testing this change, I noticed that downloading the old serf
can fail breaking the build but this is unrelated to this patch.

Authors:	Don Lewis, maho@
Obtained from:	FreeBSD port


git-svn-id: https://svn.apache.org/repos/asf/openoffice/trunk/main@1654282 13f79535-47bb-0310-9956-ffa450edef68