Re-implement Calc's RAND() function using a variant of KISS PRNG.
George Marsaglia's KISS algorithm is a rather simple implementation
of a random number generator but it has interesting properties with 
respect ot the traditional Mersenne Twister.

I used the 2011 32 bit Multiply-with-carry variant, with an
undetermined  period which is known to be not less than 10^40000000;
it is known to pass all the statistical tests.

The Apache OpenOffice implementation uses an aggressive seeding
scheme based on the internal rtl_random functions, rendering the
function basically unpredictable. As a side effect it is also not
possible to specify a seed for repeatability. No claims are made 
concerning crypto-safeness.

The specific adaption for Apache OpenOffice is:

Copyright 2015 Pedro Giffuni
All rights reserved.

Huge thanks to Steve Kargl for pointing me to the algorithm and
the late George Marsaglia for creating it in the first place.


git-svn-id: https://svn.apache.org/repos/asf/openoffice/trunk/main@1669455 13f79535-47bb-0310-9956-ffa450edef68