Bring to SAL a wrapper implementation of the powr() IEEE 754 math function.
This function was originally brought in r1444110 and is useful for
OOXML compatibility. At this time it is not being used to fix the
original issue it was meant to address but it is convenient for
future use.

It also reduces the differences from FreeBSD's port which will
be receiving updates.



git-svn-id: https://svn.apache.org/repos/asf/openoffice/trunk/main@1669456 13f79535-47bb-0310-9956-ffa450edef68