#i126604# test framework on *BSD calls "ps" wrongly, can't find AOO to kill it if hung
On *BSD, call "ps" with "-wweo", as the "ww" is needed to produce all the columns in the output.
Without it, output is trimmed short, and the test framework can't find AOO to kill it,
leaving AOO running and causing further tests to fail.

git-svn-id: https://svn.apache.org/repos/asf/openoffice/trunk@1710472 13f79535-47bb-0310-9956-ffa450edef68
(cherry picked from commit 8716daac47832a8aecd6ba509a7944c5ea3edb11)