#i35093# Find (and replace) on selected cells is misleading
When any cellc in Calc are selected at the time the "Find & Replace" dialog is opened,
automatically check the "Current selection only" checkbox, just like Writer does when
multiple lines of text are selected when this dialog opens. Users expect this from
Writer and other spreadsheets, and not doing it can even lead to wrong cells
being changed when "Replace" is used!

Patch by: me



git-svn-id: https://svn.apache.org/repos/asf/openoffice/trunk@1714091 13f79535-47bb-0310-9956-ffa450edef68