Bring new crosextra font: Caladea
Caladea is a modern Latin typeface inspired by the contrast, style and
ornaments of Khmer typefaces and writing styles. Its main objective is to
be used to write Latin texts in a Khmer context, but it is also an elegant
choice for all kinds of texts.

Designed by Carolina Giovagnoli and Andres Torresi
for Huerta Tipografica under an Apache 2.0 License.
____

This is specially interesting as it is metrically compatible with
MS Cambria.

While here sort out some issues:
- Sadly, we don't really carry an equivalent for "Arial Narrow".
- There is not need for Arimo in the Windows platform.


git-svn-id: https://svn.apache.org/repos/asf/openoffice/trunk/main@1721776 13f79535-47bb-0310-9956-ffa450edef68