Optionally bring new crosextra font: Carlito
Carlito is a modern, friendly sans-serif font, metric-compatible with 
Microsoft's Calibri font. Carlito comes in regular, bold, italic, and bold 
italic. The  family covers Latin-Greek-Cyrillic (not a complete set, 
though) with about 2,000 glyphs.

Designed by Lukasz Dziedzic available under Open Font License.

____

In general it's a bad practice to bring new features/fonts under
a Category B license: in this case we do it purely for compatibility
with other office suites and we apply the regular rules for Category B.


git-svn-id: https://svn.apache.org/repos/asf/openoffice/trunk@1722202 13f79535-47bb-0310-9956-ffa450edef68