Attempt to update libpng
libpng is is a dependency for cairo, but cairo is rarely used within
AOO and if required it is usually brought in through the system cairo.
We don't even carry libpng in the dependency list.

Updating this library is mostly innocuous but considering
CVE-2015-8540 it is convenient to keep our copy updated JIC.


git-svn-id: https://svn.apache.org/repos/asf/openoffice/trunk/main@1722371 13f79535-47bb-0310-9956-ffa450edef68