#i118023# Calc: Cut-and-paste between spreadsheets causes incorrect cell reference changes
When pasting cut cells, Calc updates references to their old positions
to instead refer to their new positions. This is done using the tab index,
row and column, however the document is not taken into account. As a result,
when cutting and pasting between documents, cells in the target document
end up getting changed instead of in the source, potentially leading to
formula corruption, which is undoable but could easily go unnoticed,
causing data loss when the document is saved.

We don't really support inter-document reference updates anyway, so fix
this bug by restricting reference updates to the intra-document cut and
paste case only.

Patch by: me
Reviewed by: kschenk



git-svn-id: https://svn.apache.org/repos/asf/openoffice/trunk/main@1730772 13f79535-47bb-0310-9956-ffa450edef68