As per https://support.microsoft.com/en-us/kb/140584 on Windows, MSVC's compiler (cl.exe) takes command-line parameters such as /MD[d] and /MT[d] to determine whether to link the C runtime library: * statically or dynamically * singly or multi threaded * debug or release
All that /MD[d] and /MT[d] do is define or undefine the _MT,
_DLL and _DEBUG macros that do the real work, and the compiler then
uses those macros to decide what to do when building the object
file.

Our build systems do not use the compiler command-line
parameters, but define the macros directly.

Of our 2 build systems, dmake always defines _MT and _DLL to get
MSVCRT.DLL linked dynamically, but while gbuild defines _MT for
all binary types, it only defines _DLL for dynamic libraries. This
means executables trying to use the C++ standard library fail to
link, as they get multiply defined symbols!!! This was first caught
in basegfx GoogleTests which #include <iostream> and create classes
from it.

This patch changes Windows and OS/2 to define  _DLL_ for dynamic
libaries instead of _DLL as that has a special meaning, and define
_DLL for all binary types on Windows so MSVCRT.DLL is always linked
dynamically like it is with dmake.

With this patch, Windows finally builds.

Patch by: me



git-svn-id: https://svn.apache.org/repos/asf/openoffice/branches/gbuild-reintegration@1736221 13f79535-47bb-0310-9956-ffa450edef68