#i126890# code: compiling with recent clang generates thousands of infinite-recu rsion warnings about DbgOut()
If the definition of the overloaded function DbgOut() in
sal/inc/rtl/string.hxx is compiled without having first compiling
the the declaration of DbgOut() with a different signature in
tools/inc/tools/debug.hxx, the compiler thinks DbgOut() is calling
itself recursively with no means of escape.
 
Fix this in the case where DBG_UTIL is defined by including
<tools/debug.hxx> inside string.hxx so that the compiler will 
encounter the declaration of DbgOut() in debug.hxx first in this
case.  This requires adding a couple of directories to the compiler
include path so that the new consumers of <tools/debug.hxx> can 
find it as well as one of its dependencies.

If DBG_UTIL is not defined, then debug.hxx does not declare DbgOut(),
so fix this case by hiding the DbgOut() definition in string.hxx
behind #ifdef DBG_UTIL since it won't be used in this case.



git-svn-id: https://svn.apache.org/repos/asf/openoffice/trunk@1736611 13f79535-47bb-0310-9956-ffa450edef68