Give up on using Perl's LWP::UserAgent and LWP::Protocol::https to download files during ./bootstrap. It's a nightmare to get it working on the buildbots - Infra has been trying on INFRA-11296 for 5 months to get it installed. Installing through CPAN doesn't always work - tests fail on CentOS 5 and on Cygwin. Even when installed, it's not always found. The gain just doesn't justify the effort. Worst of all, it's holding back development.
What else is there? A CLI tool like wget could work, but it's not listed as
a dependency, and it breaks on CentOS 5 for https://, the thing it's needed
for most.

I instead re-implemented it in Java. Java is freely available, highly
portable, and rock solid. We already use it in the build, and it's
described as being a mandatory build requirement even though
./configure.ac treats it as optional. Best of all it supports https://
out of the box in java.net.URLConnection and uses its own root CA
certificates. Tests show my AOOJavaDownloader class works on FreeBSD and
Windows, supports HTTP redirection, and generally works like a charm.

Patch by: me



git-svn-id: https://svn.apache.org/repos/asf/openoffice/trunk@1753943 13f79535-47bb-0310-9956-ffa450edef68