apr has 3 different ways of building. By default, if devenv is in the $PATH, it is used. If not, and apr.mak is found, nmake is used. Failing both, msdev is used.
On the buildbots, devenv is somehow in the $PATH, even when
VCVARS32.BAT is not run, causing the first way to be
used. This causes some kind of deadlock that hangs the
build and prevents deletion of Makefile.win until devenv
is killed.

We know the nmake way works, so force it to be used instead.

Patch by: me



git-svn-id: https://svn.apache.org/repos/asf/openoffice/trunk@1754057 13f79535-47bb-0310-9956-ffa450edef68