#i126896#:  bundled curl version 7.19.7 has many vulnerabilities
Upgrade bundled curl to version 7.49.1.

Delete most of the curl patches since they are no longer necessary.
The only needed patch is to produce a library with the expected name
when building on Windows.

Update the curl copyright info in LICENSE.
 
These curl CVEs have been fixed since 7.19.7:
    CVE-2010-0734
    CVE-2011-2192
    CVE-2013-2174
    CVE-2014-3143
    CVE-2014-3144
    CVE-2014-3145
    CVE-2014-3148
    CVE-2014-8150
    CVE-2015-3153
    CVE-2016-0755
Whether any of these affect the OpenOffice usage of curl is not known. 
OpenOffice only uses curl to access ftp:// URLs, which is likely to be
only rarely done.



git-svn-id: https://svn.apache.org/repos/asf/openoffice/trunk@1754469 13f79535-47bb-0310-9956-ffa450edef68