#i117845#  [gbuild] invalid link-target dep-files can be generated
Fix the Windows build performance regression with precompiled headers
introduced by the fix for #i117845# that was merged from
branches/gbuild-reintegration by r1755455.

The use of precompiled headers is triggered by the use of two compiler
flags, -DPRECOMPILED_HEADERS which triggers the inclusion of a special
header for the library that includes all of the include files needed
for any source file in the library, and -Yu[headername] which tells the
the Microsoft compiler to use the precompiled version of this header.

It would be nice if these two flags were added to the compiler flags
at the same time, but they are not.  The -DPRECOMPILED_HEADERS flag
is always added if the library is configured to use precompiled
headers, but the -Yu flag is only added conditionally if the flags
for building  a particular source file match those that were used to
precompile the header.  A further condition is the presence of
$(PCH_NAME), which is set as a target-specific variable on the
library, and which is supposed to be inherited by its dependencies
to be used when compiling the source files for the library.
Unfortunately the fix for #i117845 caused the source and object
files to be dependencies of a different target, which did not
have $(PCH_NAME) set.  They did not inherit $(PCH_NAME), so they
were compiled with -DPRECOMPILED_HEADERS, but not -Yu, causing the
compiler to do more work than in the case where the precompiled
header option is disabled resulting in longer builds with the  use
of this option.

Fix by setting a target-specific $(PCH_NAME) for the new target.



git-svn-id: https://svn.apache.org/repos/asf/openoffice/trunk@1758061 13f79535-47bb-0310-9956-ffa450edef68