Update Python to version 2.7.12.
This is the latest bugfix release in the 2.7.x series (Released 2016-06-28).

We expect this to be the best release that still guarantees some 
compatibility with previous versions while still letting us use some of 
the features in newer language versions.

It is also easier to keep up to date with upstream and with newer 
compilers.

Tested on FreeBSD, but shouldn't have problems on other platforms.


git-svn-id: https://svn.apache.org/repos/asf/openoffice/trunk@1759046 13f79535-47bb-0310-9956-ffa450edef68