Adapt TrueType font detection for the Apple specification variant.
From the Microsoft OpenType Specification:
https://www.microsoft.com/typography/otspec/otff.htm

NOTE: The Apple specification for TrueType fonts allows for 'true' and 
'typ1' for sfnt version. These version tags should not be used for fonts 
which contain OpenType tables.
____

We are complying with the OpenType usage fine but we were not taking into 
account the NOTE so Apple TrueType variants would not be recognized.

Apparently valid values are 0x00010000 and 0x00020000. Keep it simple and
just assume any value different than zero is valid here.


git-svn-id: https://svn.apache.org/repos/asf/openoffice/trunk@1765086 13f79535-47bb-0310-9956-ffa450edef68