#i127165# Clicking "No" in Java dialog causes abnormal exit on FreeBSD
XSingleComponentFactory::createInstanceWithArgumentsAndContext()
throws an Exception, not a RuntimeException. This was causing
an abort on FreeBSD when "No" is clicked in the Java dialog,
as JavaComponentLoader::activate() throws a
CannotActivateFactoryException which is a subclass of Exception
but not RuntimeException, and the scripting ProviderCache wrongly
catches only RuntimeException. The missed CannotActivateFactoryException
was going to unexpected(), which calls abort()...

Patch by: me



git-svn-id: https://svn.apache.org/repos/asf/openoffice/trunk@1765170 13f79535-47bb-0310-9956-ffa450edef68