#i83004# cannot paste HTML data from any Java application via clipboard
Allow StartHTML and EndHTML values in the Windows clipboard's
"HTML Format" to be -1, and use StartFragment and EndFragment instead
when they are. Excel allows -1, as does Mozilla since 2011, so we
really should too. Java has been providing >= 0 for a while now, but
we should still support -1 in case any other applications use it.

Patch by: me



git-svn-id: https://svn.apache.org/repos/asf/openoffice/trunk@1765837 13f79535-47bb-0310-9956-ffa450edef68