On Linux, building without --enable-dbgutil causes linking failures in GoogleTests in main/binaryurp with missing pthread symbols.
The original cause is that ext_libraries/test's libgtest.so is
built without libpthread.so linked, which is what this commit fixes
(or rather hacks around; not sure why that happens).

There must also be a proximal cause in main/binaryurp's tests,
as they link with --enable-dbgutil even without this commit, and
can be made to link even without --enable-dbgutil by changing the
OSL_DEBUG_LEVEL define in the compiler options from 0 to 1 (this
doesn't however cause any changes in the preprocessed source file,
so I haven't been able to debug it further).

This should hopefully get the Linux buildbots working again.

Patch by: me



git-svn-id: https://svn.apache.org/repos/asf/openoffice/trunk@1776043 13f79535-47bb-0310-9956-ffa450edef68