#i127461#: Update bundled expat to version 2.2.1
Version 2.2.0 that is bundled in trunk has vulnerabilities CVE-2017-9233
and CVE-2016-9063 and other potential problems.  It is not known whether
these impact OpenOffice.

The patch for <https://sourceforge.net/p/expat/bugs/539/> is included so
we no longer need a local patch for that.  We do need a new patch to
work around the lack of <stdint.h> in MS Visual Studio 9.0.



git-svn-id: https://svn.apache.org/repos/asf/openoffice/trunk@1800568 13f79535-47bb-0310-9956-ffa450edef68