Be slightly less tricky and change 	do { 	} while (!this); to 	do { 	} while (0);
Computing !this must be done at runtime, whereas the intent is for
the loop to exit after the first iteration, which can be handled at
compile time when using a constant zero.  Also, "this" should always
be non-NULL in valid c++ code, but if ever happened to be NULL, then
this code could loop forever.
 
While here, translate nearby comments.



git-svn-id: https://svn.apache.org/repos/asf/openoffice/trunk@1822695 13f79535-47bb-0310-9956-ffa450edef68